Provide execution context for document resolution.
Let the phoenix channel provide execution context if its available
on the request payload. This way resolve functions can have access
to context values (in adition to the document variables), for
example, some middleware can be set to send always send the user
authorization token to provide a context for the query being
executed.

Tests implement a simple `viewer` query that resolve to a user
if provided a "jwt_token" value on context.