Fix `Absinthe.Phoenix.Controller` documentation.
If one doesn't provide the `action: [mode: :internal]` addendum to the
`use` statement then Absinthe returns results serialised as if it were
about to convert it to JSON, rather than returning the unserialized
data.  Most importantly this means that map keys are atoms rather than
strings.

Thanks to @aspett for the tip.